+++ apps/learn-card-app/src/components/boost/boostCMS/boostCMSForms/boostCMSIssueTo/BoostAddressBookContactOptions.tsx
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import { Capacitor } from '@capacitor/core';
 import { Share } from '@capacitor/share';
 import { Clipboard } from '@capacitor/clipboard';
-import { useLocation, useHistory } from 'react-router-dom';
 import { BarcodeScanner } from '@capacitor-mlkit/barcode-scanning';
 
 import {
@@ -35,7 +34,6 @@ import { ModalTypes, QRCodeScannerStore, useModal, useToast, ToastTypeEnum } fro
 import { BoostCMSIssueTo, BoostCMSState } from '../../../boost';
 import useTheme from '../../../../../theme/hooks/useTheme';
 import { useWallet } from 'learn-card-base';
-import { usePathQuery } from 'learn-card-base';
 import { useGetCurrentLCNUser } from 'learn-card-base';
 
 import { BoostUserTypeEnum } from 'learn-card-base';
@@ -86,9 +84,6 @@ const BoostAddressBookContactOptions: React.FC<BoostAddressBookContactOptionsPro
     const { currentLCNUser, currentLCNUserLoading } = useGetCurrentLCNUser();
     const [_issueTo, _setIssueTo] = useState<BoostCMSIssueTo[]>(state?.[collectionPropName]);
     const [walletDid, setWalletDid] = useState<string>('');
-    const query = usePathQuery();
-    const history = useHistory();
-    const location = useLocation();
 
     useEffect(() => {
         const getWalletDid = async () => {
@@ -264,16 +259,6 @@ const BoostAddressBookContactOptions: React.FC<BoostAddressBookContactOptionsPro
     const handleAddSomeoneElse = () => {
         setIssueMode?.(BoostUserTypeEnum.someone);
 
-        if (history) {
-            let searchParams = new URLSearchParams(window.location.search);
-            searchParams.set('boostUserType', BoostUserTypeEnum.someone);
-
-            history.replace({
-                pathname: location?.pathname,
-                search: searchParams.toString(),
-            });
-        }
-
         presentAddressBook();
         // newModal(
         //     <BoostAddressBook
@@ -325,7 +310,11 @@ const BoostAddressBookContactOptions: React.FC<BoostAddressBookContactOptionsPro
                 <IonGrid className="ion-padding">
                     <IonRow className="w-full flex items-center justify-center mt-8">
                         <button
-                            onClick={() => handleAddYoself()}
+                            onClick={event => {
+                                event.preventDefault();
+                                event.stopPropagation();
+                                handleAddYoself();
+                            }}
                             className={`${
                                 primaryColor ? `bg-${primaryColor}` : 'bg-grayscale-900'
                             } text-xl text-white flex items-center justify-center font-semibold py-[5px] rounded-full w-full border-solid border-white border-[2px] px-[18px] shadow-soft-bottom`}
@@ -336,7 +325,11 @@ const BoostAddressBookContactOptions: React.FC<BoostAddressBookContactOptionsPro
                     </IonRow>
                     <IonRow className="w-full flex items-center justify-center mt-4">
                         <button
-                            onClick={() => handleAddSomeoneElse()}
+                            onClick={event => {
+                                event.preventDefault();
+                                event.stopPropagation();
+                                handleAddSomeoneElse();
+                            }}
                             className={`${
                                 primaryColor ? `bg-${primaryColor}` : 'bg-grayscale-900'
                             } text-xl text-white flex items-center justify-center font-semibold py-[5px] rounded-full w-full border-solid border-white border-[2px] px-[18px] shadow-soft-bottom`}
@@ -375,7 +368,11 @@ const BoostAddressBookContactOptions: React.FC<BoostAddressBookContactOptionsPro
                     {Capacitor.isNativePlatform() && (
                         <IonCol className="w-full flex items-center justify-center mt-1">
                             <button
-                                onClick={handleScan}
+                                onClick={event => {
+                                    event.preventDefault();
+                                    event.stopPropagation();
+                                    handleScan();
+                                }}
                                 className="flex items-center justify-center bg-grayscale-900 rounded-full px-[18px] py-[12px] text-white font-poppins text-xl w-full shadow-lg modal-btn-mobile normal tracking-wide"
                             >
                                 <Camera className="ml-[5px] h-[30px] w-[30px] mr-2" /> Scan Code